Corrected some scripts

+++ hg38ref_run/run_scripts/run_MAF_speciesLister.sh
@@ -3,10 +3,10 @@
 #SBATCH -A snic2021-5-28
 #SBATCH -p core
 #SBATCH -n 4
-#SBATCH -e error_out/speciesLister/branch241_%j_%a.err
 #SBATCH -J branch241_%a
+#SBATCH -e error_out/speciesLister/branch241_%j_%a.err
 #SBATCH -o error_out/speciesLister/branch241_%j_%a.out
-#SBATCH -t 1-00:00:00
+#SBATCH -t 12:00:00
 
 echo -n "Time started: "
 date
@@ -34,7 +34,19 @@ fname2=${fname%.*}
 mafFILE=${FILE%.*}
 
 # uncompress target file, keep original
-/proj/uppstore2017228/KLT.04.200M/200m_MD/resources/zstd/zstd-dev/zstd -d -k $FILE
+# need to install zstd if you use zstd-compressed files
+dontdelete=0
+if [[ "$fnamebase" =~ ^*.gz$ ]]; then
+        gzip -d -c $FILE > $mafFILE
+elif [[ "$fnamebase" =~ ^*.zst$ ]]; then
+        zstd -d -k $FILE
+elif [[ "$fnamebase" =~ ^*.maf$ ]]; then
+		fnamebase=${FILE##*/}
+		fname=${fnamebase}
+		fname2=${fname%.*}
+		mafFILE=$FILE
+		dontdelete=1 # important
+fi
 
 output_dir=$2
 output_file=$output_dir"/"$fname2".bed"
@@ -46,7 +58,9 @@ python MAF_allele_stats_get_names_v1.py $mafFILE $output_file $reference_species
 gzip $output_file
 
 # Delete decompressed file and temp files
+if [ $dontdelete == 0 ]; then
 rm $mafFILE
+fi
 
 echo -n "Time ended: "
 date
